+++ util/iron/CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## Unreleased
 ### Added
 - Advanced `Router`.
+- Sentry error response middlewere.
 
 ## 0.1.3 - 2019-03-29
 ### Added